+++ imgboard.php
@@ -1,5 +1,5 @@
 <?php
-define('NELLIEL_VERSION', 'v0.9.6.14'); // Version
+define('NELLIEL_VERSION', 'v0.9.6.15'); // Version
 define('NELLIEL_COPYRIGHT', '2010-2018 Nelliel Project'); // Copyright line
 define('NELLIEL_PACKAGE', 'Nelliel'); // Package
 define('BASE_PATH', realpath('./') . '/'); // Base path for script
@@ -34,11 +34,11 @@ require_once INCLUDE_PATH . 'crypt.php';
 
 nel_set_password_algorithm(NEL_PASSWORD_PREFERRED_ALGORITHM);
 
-require_once INCLUDE_PATH . 'setup/setup.php';
-
 if (RUN_SETUP_CHECK)
 {
-    setup_check(INPUT_BOARD_ID);
+    $setup = new \Nelliel\setup\Setup();
+    $board_id = (isset($_GET['board_id'])) ? $_GET['board_id'] : '';
+    $setup->checkAll($board_id);
 }
 
 require_once INCLUDE_PATH . 'regen.php';